requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
networkx>=3.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
